python/requirements.txt
google-cloud-pubsub
protobuf
python-dotenv
